rag/requirements.txt
numpy==1.26.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
requests==2.31.0
